+++ src/main/java/team/cqr/cqrepoured/client/util/BlockRenderUtil.java
@@ -22,7 +22,7 @@ public class BlockRenderUtil {
 		int i = currentEntity.getBrightnessForRender();
 		int j = i % 65536;
 		int k = i / 65536;
-		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, (float) j, (float) k);
+		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, j, k);
 		GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
 		renderer.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
 		blockrendererdispatcher.renderBlockBrightness(iBlockState, 1.0F);
