+++ src/main/java/team/cqr/cqrepoured/crafting/RecipesArmorDyes.java
@@ -57,13 +57,13 @@ public class RecipesArmorDyes extends IForgeRegistryEntry.Impl<IRecipe> implemen
 
 					if (itemarmor.hasColor(itemstack1)) {
 						int l = itemarmor.getColor(itemstack);
-						float f = (float) (l >> 16 & 255) / 255.0F;
-						float f1 = (float) (l >> 8 & 255) / 255.0F;
-						float f2 = (float) (l & 255) / 255.0F;
-						i = (int) ((float) i + Math.max(f, Math.max(f1, f2)) * 255.0F);
-						aint[0] = (int) ((float) aint[0] + f * 255.0F);
-						aint[1] = (int) ((float) aint[1] + f1 * 255.0F);
-						aint[2] = (int) ((float) aint[2] + f2 * 255.0F);
+						float f = (l >> 16 & 255) / 255.0F;
+						float f1 = (l >> 8 & 255) / 255.0F;
+						float f2 = (l & 255) / 255.0F;
+						i = (int) (i + Math.max(f, Math.max(f1, f2)) * 255.0F);
+						aint[0] = (int) (aint[0] + f * 255.0F);
+						aint[1] = (int) (aint[1] + f1 * 255.0F);
+						aint[2] = (int) (aint[2] + f2 * 255.0F);
 						++j;
 					}
 				} else if (DyeUtils.isDye(itemstack1)) {
@@ -89,10 +89,10 @@ public class RecipesArmorDyes extends IForgeRegistryEntry.Impl<IRecipe> implemen
 			int j1 = aint[1] / j;
 			int k1 = aint[2] / j;
 			float f3 = (float) i / (float) j;
-			float f4 = (float) Math.max(i1, Math.max(j1, k1));
-			i1 = (int) ((float) i1 * f3 / f4);
-			j1 = (int) ((float) j1 * f3 / f4);
-			k1 = (int) ((float) k1 * f3 / f4);
+			float f4 = Math.max(i1, Math.max(j1, k1));
+			i1 = (int) (i1 * f3 / f4);
+			j1 = (int) (j1 * f3 / f4);
+			k1 = (int) (k1 * f3 / f4);
 			int k2 = (i1 << 8) + j1;
 			k2 = (k2 << 8) + k1;
 			itemarmor.setColor(itemstack, k2);
