+++ src/main/java/team/cqr/cqrepoured/objects/entity/ai/boss/endercalamity/BossAIEndLaser.java
@@ -30,10 +30,11 @@ public class BossAIEndLaser extends AbstractBossAIEnderCalamity {
 		if (this.entity.hasHomePositionCQR()) {
 			BlockPos home = this.entity.getHomePositionCQR();
 			this.entity.teleport(home.getX(), home.getY(), home.getZ());
-			//TODO: Remove blocks in the center
+			// TODO: Remove blocks in the center
 		}
-		
-		// IMessage message = SPacketUpdateAnimationOfEntity.builder(this.entity).animate(EntityCQREnderCalamity.ANIM_NAME_LASER_STATIONARY).build();
+
+		// IMessage message =
+		// SPacketUpdateAnimationOfEntity.builder(this.entity).animate(EntityCQREnderCalamity.ANIM_NAME_LASER_STATIONARY).build();
 		// CQRMain.NETWORK.sendToAllTracking(message, this.entity);
 		this.entity.sendAnimationUpdate(EntityCQREnderCalamity.ANIM_NAME_LASER_STATIONARY);
 	}
@@ -50,11 +51,11 @@ public class BossAIEndLaser extends AbstractBossAIEnderCalamity {
 		}
 		if (this.endlaser != null) {
 			// DONE: Fix buggy rotation
-			this.entity.rotationYaw = (float) this.endlaser.rotationYawCQR + 90.0F;
-			this.entity.prevRotationYaw = (float) this.endlaser.prevRotationYawCQR + 90.0F;
+			this.entity.rotationYaw = this.endlaser.rotationYawCQR + 90.0F;
+			this.entity.prevRotationYaw = this.endlaser.prevRotationYawCQR + 90.0F;
 
-			this.entity.rotationPitchCQR = (float) this.endlaser.rotationPitchCQR;
-			this.entity.prevRotationPitchCQR = (float) this.endlaser.prevRotationPitchCQR;
+			this.entity.rotationPitchCQR = this.endlaser.rotationPitchCQR;
+			this.entity.prevRotationPitchCQR = this.endlaser.prevRotationPitchCQR;
 
 			this.entity.rotationYawHead = this.entity.rotationYaw;
 			this.entity.prevRotationYawHead = this.entity.prevRotationYaw;
@@ -69,7 +70,8 @@ public class BossAIEndLaser extends AbstractBossAIEnderCalamity {
 		// Head distance with scale = 100%: 0.75 blocks
 		float yaw = this.entity.rotationYaw;
 		if (this.entity.hasAttackTarget()) {
-			yaw = (float) Math.toDegrees(Math.atan2(-(this.entity.getAttackTarget().posX - this.entity.posX), this.entity.getAttackTarget().posZ - this.entity.posZ));
+			yaw = (float) Math
+					.toDegrees(Math.atan2(-(this.entity.getAttackTarget().posX - this.entity.posX), this.entity.getAttackTarget().posZ - this.entity.posZ));
 		}
 
 		AbstractEntityLaser laser = new EntityEndLaser(this.entity.getEntityWorld(), this.entity, 64.0F, 8.0F, -0.01F);
@@ -86,7 +88,8 @@ public class BossAIEndLaser extends AbstractBossAIEnderCalamity {
 			this.endlaser.setDead();
 			this.endlaser = null;
 		}
-		// IMessage message = SPacketUpdateAnimationOfEntity.builder(this.entity).animate(EntityCQREnderCalamity.ANIM_NAME_IDLE_BODY).build();
+		// IMessage message =
+		// SPacketUpdateAnimationOfEntity.builder(this.entity).animate(EntityCQREnderCalamity.ANIM_NAME_IDLE_BODY).build();
 		// CQRMain.NETWORK.sendToAllTracking(message, this.entity);
 		this.entity.sendAnimationUpdate(EntityCQREnderCalamity.ANIM_NAME_IDLE_BODY);
 		this.entity.setCantUpdatePhase(false);
